Fix breakpoint binding for anonymous methods.
We consider a breakpoint to be invalid and stop looking for an IL
location when the break line is inside the block of lines of a method,
but there is no IL location for that line. But that's not always
invalid. For example, the block of a method may contain code from an
anonymous method. So, even if a breakpoint looks invalid, we have to
keep looking, since it may be a valid location of an anon method.