Improved compatibility with VS/SD file formats
This is a partial fix of bug #664644. MD will now generate project
configurations for all config/platform combinations found in the
project file, even when there isn't a property group for every
configuration/platform. So for example, if the msbuild file has a
property group for Debug and another one for x86, this will generate a
Debug|x86 configuration. When saving the project, this new
configuration will not be saved, unless configuration specific values
have been set. Also, when merging/unmerging properties, do case
insensitive comparison of values (msbuild is supposed to be case
insensitive).