Optimized memory usage
When parsing a big code base (such as a clean checkout of MD), there
may be many type resolution failures since many types are not
registered until a whole parse of all projects is finished. When there
is a type resolution failure in a file, the file is queued for parsing
again. Before this patch, the parser db wasn't actually queuing the
file for reparse, but it was queuing a job for retrying the type
resolution on the already parsed compilation unit object. For a big
code base as MD, it ended queuing thousands of compilation units,
which take big amounts of memory. To avoid this problem, we now queue
a new reparse request and we don't keep the compilation unit in memory
anymore. However, we only do it if the original parse request was done
for the on-disk file content, otherwise we use the old behavior.