[MacDev] Don't write non-partial methods as comments
For non-C# files, we used to generate designer methods
as text, and put them in comments. Now, we would strip
out such comments when round-tripping to Obj-C, so
instead write real methods. The user can move it to
a non-designer partial class manually, and it would not be 
generated in the designer part again.